Changed the import reference for searchhits after the move
