+++ src/main/java/dk/sdu/se_f22/sortingmodule/range/InternalFilter.java
@@ -1,6 +1,6 @@
 package dk.sdu.se_f22.sortingmodule.range;
 
-import dk.sdu.se_f22.sortingmodule.infrastructure.SearchHits;
+import dk.sdu.se_f22.sharedlibrary.SearchHits;
 
 import java.util.ArrayList;
 import java.util.Collection;
