+++ src/main/java/dk/sdu/se_f22/sortingmodule/range/RangeFilterInterface.java
@@ -1,6 +1,6 @@
 package dk.sdu.se_f22.sortingmodule.range;
 
-import dk.sdu.se_f22.sortingmodule.infrastructure.SearchHits;
+import dk.sdu.se_f22.sharedlibrary.SearchHits;
 import dk.sdu.se_f22.sortingmodule.range.exceptions.InvalidFilterIdException;
 
 import java.util.List;
